\name{EppsPulley.test}
\alias{EppsPulley.test}
\title{Approximate Epps-Pulley normality test}
\usage{
EppsPulley.test(x, alpha = 0.05)
}
\arguments{
  \item{x}{Sample}

  \item{alpha}{Tolerance level}
}
\value{
A list containing the test statistics
}
\description{
An Approximation to the Limit Distribution of the
Epps-Pulley Test Statistic for Normality By N. Henze
Metrika (1990) 37:7-18
}
\author{
François Pelletier
}

