\name{zerobond}
\alias{zerobond}
\title{zerobond}
\usage{
zerobond(eval.time, expiry.time, rate, face = 1)
}
\arguments{
  \item{eval.time}{Evaluation time}

  \item{expiry.time}{Expiry time}

  \item{rate}{Continuously compounded interest rate (force
  of interest)}

  \item{face}{Face value}
}
\value{
Actualised price of bond
}
\description{
Evaluate the price of a zero coupon bond
}
\details{
Evaluates the actualised price of a bond using interest
rate and face value
}
\author{
François Pelletier
}

